init_encrypted_tunnel -> init_encrypted_tunnels, base_url to utils, formdata functions collapsed
