+++ src/fetch/formdata.rs
@@ -1,25 +1,22 @@
 use js_sys::Uint8Array;
 use wasm_bindgen::{JsCast, JsValue};
 
-// Converts an instance of `web_sys::FormData` to a `Uint8Array`
+/// Converts an instance of `web_sys::FormData` to a `Uint8Array`
+///
+///  // Sample output:
+/// --AaB03x
+/// content-disposition: form-data; name="field1"
+/// content-type: text/plain;charset=windows-1250
+/// content-transfer-encoding: quoted-printable
+///
+/// Joe owes =80100.
+/// --AaB03x
+///
+/// Ref: <https://github.com/nodejs/undici/blob/e39a6324c4474c6614cac98b8668e3d036aa6b18/lib/fetch/body.js#L31>
 pub async fn parse_form_data_to_array(
     form: web_sys::FormData,
-    boundary: String,
+    boundary: &str,
 ) -> Result<Vec<u8>, JsValue> {
-    // Sample output:
-    //
-    //     --AaB03x
-    //     content-disposition: form-data; name="field1"
-    //     content-type: text/plain;charset=windows-1250
-    //     content-transfer-encoding: quoted-printable
-    //
-    //     Joe owes =80100.
-    //     --AaB03x
-    extract_body(form, &boundary).await
-}
-
-// Ref: <https://github.com/nodejs/undici/blob/e39a6324c4474c6614cac98b8668e3d036aa6b18/lib/fetch/body.js#L31>
-async fn extract_body(form: web_sys::FormData, boundary: &str) -> Result<Vec<u8>, JsValue> {
     let prefix = format!("--{}\r\nContent-Disposition: form-data", boundary);
     let mut blob_parts = Vec::new();
     let rn = Uint8Array::from(&[13, 10][..]); // '\r\n'
@@ -98,6 +95,7 @@ fn escape(str: &str) -> String {
         .replace('\r', "%0D")
         .replace('"', "%22")
 }
+
 fn normalize_linefeeds(value: &str) -> String {
     value.replace("\r\n", "\n").replace('\r', "\n")
 }
