+++ src/utils.rs
@@ -1,3 +1,5 @@
+use wasm_bindgen::JsValue;
+
 pub(crate) async fn sleep(delay: i32) {
     let mut cb = |resolve: js_sys::Function, _: js_sys::Function| {
         _ = web_sys::window()
@@ -7,4 +9,17 @@ pub(crate) async fn sleep(delay: i32) {
 
     let p = js_sys::Promise::new(&mut cb);
     wasm_bindgen_futures::JsFuture::from(p).await.unwrap();
+}
+
+pub(crate) fn base_url(url: &str) -> Result<String, JsValue> {
+    let url =
+        url::Url::parse(url).map_err(|e| JsValue::from_str(&format!("Invalid URL: {}", e)))?;
+
+    // get without query or path fragments
+    let mut base_url = format!("{}://{}", url.scheme(), url.host_str().unwrap_or_default());
+    if let Some(port) = url.port() {
+        base_url = format!("{}:{}", base_url, port);
+    }
+
+    Ok(base_url)
 }
